refactored code: moved state out of SearchForm and up into App.
